Rename file_router.py -> file_type_router.py, add __init__.py
