+++ haystack/preview/components/routers/__init__.py
@@ -1,0 +1,4 @@
+from haystack.preview.components.routers.file_type_router import FileTypeRouter
+from haystack.preview.components.routers.metadata_router import MetadataRouter
+
+__all__ = ["FileTypeRouter", "MetadataRouter"]
